+++ macros/support/src/com_interface/com_interface_impl.rs
@@ -1,4 +1,3 @@
-use super::vtable_macro;
 use super::Interface;
 use crate::com_interface::{iid, vtable};
 
@@ -9,7 +8,7 @@ pub fn generate(interface: &Interface) -> TokenStream {
     let interface_ident = &interface.name;
     let vtable_ident = vtable::ident(&interface_ident.to_string());
     let iid_ident = iid::ident(interface_ident);
-    let vtable_macro = vtable_macro::ident(&interface_ident);
+    // let vtable_macro = vtable_macro::ident(&interface_ident);
     let parent = if let Some(p) = &interface.parent {
         quote! { #p }
     } else {
