+++ src/production.rs
@@ -1,0 +1,8 @@
+mod co_class;
+#[doc(hidden)]
+pub mod offset;
+#[doc(hidden)]
+pub mod registration;
+
+pub use co_class::CoClass;
+pub use co_class::ProductionComInterface;
